% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kom2020}
\alias{kom2020}
\title{Kommune data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 389 rows and 5 columns.
}
\usage{
data(kom2020)
}
\description{
Kommune data
}
\keyword{datasets}
